skills/windows-remote/agent/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pyautogui>=0.9.54
mss>=9.0.0
Pillow>=10.0.0
